Advanced AI, M Minimap disabled.
-AI now uses more complex behaviors. Summary:
--If an enemy sector spawns, each planet gets a defensive squad that
will chase the player if nearby, and the sector gets an offensive squad.
-The offensive squad goes to attack the nearest player planet. If it
wins the space battle, it invades. If it wins the invasion, it becomes
the defensive squad for that planet. Otherwise, it turns home and
defends the sector. If it lost all soldiers in the space battle, it
returns home.

Disabled the minimap because it was making my laptop run out of memory.
Holy shit.